Move cron schedule to worker defined
