+++ oxanus/tests/integration/cron.rs
@@ -1,6 +1,29 @@
 use crate::shared::*;
 use deadpool_redis::redis::AsyncCommands;
+use serde::{Deserialize, Serialize};
 use testresult::TestResult;
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct CronWorkerRedisCounter {}
+
+#[async_trait::async_trait]
+impl oxanus::Worker for CronWorkerRedisCounter {
+    type Context = WorkerState;
+    type Error = WorkerError;
+
+    async fn process(
+        &self,
+        oxanus::Context { ctx, .. }: &oxanus::Context<WorkerState>,
+    ) -> Result<(), WorkerError> {
+        let mut redis = ctx.redis.get().await?;
+        let _: () = redis.incr("cron:counter", 1).await?;
+        Ok(())
+    }
+
+    fn cron_schedule() -> Option<String> {
+        Some("* * * * * *".to_string())
+    }
+}
 
 #[tokio::test]
 pub async fn test_cron() -> TestResult {
@@ -16,7 +39,7 @@ pub async fn test_cron() -> TestResult {
         .namespace(random_string())
         .build_from_pool(redis_pool.clone())?;
     let config = oxanus::Config::new(&storage)
-        .register_cron_worker::<CronWorkerRedisCounter>("* * * * * *", QueueOne)
+        .register_cron_worker::<CronWorkerRedisCounter>(QueueOne)
         .exit_when_processed(2);
 
     oxanus::run(config, ctx).await?;
